Debugged iGraph trans_rate function
The trans_rate function calculates a graph's transitivity rate as the proportion of two-step paths that are also one-step paths. iGraph generates named or unnamed nodelists depending on whether node labels are supplied. When generating a network from an edgelist, netwritge generates node labels. When creating networks from adjacency matrices and lists, this is not necessarily the case. The revised function now handles either use case, named or unnamed vector of node IDs. Previously, the function assumed a named list.